requirements.txt
torch
datasets
transformers
soundfile
tqdm
librosa
faiss-cpu
editdistance
jarowinkler
gensim